Initial Version
Thanks to the contract, initial implementation of this gem was
relatively straightforward. Checking if a directory exists was a
little finnicky, but it's been a bit of a problem in all of these
cloud providers. Since they all only pretend to have a directory
structure, the prefix/delimiter pattern can be a bit difficult to get
correct.